requirement.txt.txt
Libraries use in this project.
1.pandas
2.numpy
3.scikit-learn
4.matplotlib
5.seaborn
6.jupyter